Fix mastering OOM on 5+ min tracks: cut peak RSS ~25% + length cap
The intermittent 500s on /api/master were OOM-kills: peak RSS scales with
track length and crossed the 2GB instance at ~5 min. Per-stage profiling
showed the transient shaper (not the limiter) was the peak -- its tanh
waveshaper allocated 3-4 full-length float64 temporaries at once.

- transient_shaper: in-place tanh + float32 gain/output (largest saving)
- limiter: in-place gain (up *= gain), drops the duplicate 4x buffer on
  all ~7 limiter passes per request
- lowband: keep crossover intermediate bands float32 (was float64)
- mastering_chain: loudness-ride limiter oversample 4->2 (export stays 4x,
  so the -1 dBTP true-peak guarantee is unchanged); ~30% faster
- routes: 5-min length cap (413) + del raw + serialize masters (semaphore)

Verified under docker --memory=2g (matches render Standard): a 6-min track
went from OOM-kill (exit 137) to completing at 1811MB; output quality
identical (integrated LUFS -9.1, true-peak ~-3.9 dBTP before/after).
Adds bench.py/profile_mem.py/test_cap.py for memory regression checks.

+++ backend/app/dsp/lowband.py
@@ -29,11 +29,14 @@ def split_bands(audio: np.ndarray, sr: int, crossover_hz: float = 120.0):
     sos_lp = butter(2, wn, btype="low", output="sos")
     sos_hp = butter(2, wn, btype="high", output="sos")
 
-    low = sosfilt(sos_lp, audio, axis=-1)
-    low = sosfilt(sos_lp, low, axis=-1)
-    high = sosfilt(sos_hp, audio, axis=-1)
-    high = sosfilt(sos_hp, high, axis=-1)
-    return low.astype(np.float32), high.astype(np.float32)
+    # Cast back to float32 after each sosfilt: scipy upcasts to float64
+    # internally, so without this the two full-length intermediate bands are
+    # 8-byte buffers (double the RAM) until the final cast.
+    low = sosfilt(sos_lp, audio, axis=-1).astype(np.float32, copy=False)
+    low = sosfilt(sos_lp, low, axis=-1).astype(np.float32, copy=False)
+    high = sosfilt(sos_hp, audio, axis=-1).astype(np.float32, copy=False)
+    high = sosfilt(sos_hp, high, axis=-1).astype(np.float32, copy=False)
+    return low, high
 
 
 def compress_low_band(
